refactor(backend): extract sales order query service
