+++ backend/src/modules/sales/services/sales-order.service.spec.ts
@@ -21,6 +21,7 @@ import { ConflictException, NotFoundException } from '@nestjs/common';
 import { SalesOrderFulfillmentService } from './sales-order-fulfillment.service';
 import { SalesOrderLifecycleService } from './sales-order-lifecycle.service';
 import { SalesOrderPaymentService } from './sales-order-payment.service';
+import { SalesOrderQueryService } from './sales-order-query.service';
 
 describe('SalesOrderService', () => {
   let service: SalesOrderService;
@@ -61,13 +62,8 @@ describe('SalesOrderService', () => {
   });
 
   beforeEach(async () => {
-    // Mock DataSource for findById method
-    const mockPaymentRepository = {
-      find: jest.fn().mockResolvedValue([]),
-    };
-
     dataSource = {
-      getRepository: jest.fn().mockReturnValue(mockPaymentRepository),
+      transaction: jest.fn(),
     } as any;
 
     const module: TestingModule = await Test.createTestingModule({
@@ -76,6 +72,7 @@ describe('SalesOrderService', () => {
         SalesOrderFulfillmentService,
         SalesOrderLifecycleService,
         SalesOrderPaymentService,
+        SalesOrderQueryService,
         {
           provide: getRepositoryToken(SalesOrder),
           useValue: {
@@ -114,6 +111,12 @@ describe('SalesOrderService', () => {
         {
           provide: getRepositoryToken(PriceListItem),
           useValue: {},
+        },
+        {
+          provide: getRepositoryToken(Payment),
+          useValue: {
+            find: jest.fn().mockResolvedValue([]),
+          },
         },
         {
           provide: getDataSourceToken(),
